Hotfix Unit
Unitのメンバ関数の戻り値が不正であるのを修正
その他軽微な修正

+++ NutLib/Unit/Unit.hpp
@@ -52,25 +52,25 @@ public:
 	constexpr float f()const{
 		return static_cast<float>(_value);
 	}
-	constexpr float d()const{
+	constexpr double d()const{
 		return static_cast<double>(_value);
 	}
-	constexpr float i8()const{
+	constexpr int8_t i8()const{
 		return static_cast<int8_t>(_value);
 	}
-	constexpr float i16()const{
+	constexpr int16_t i16()const{
 		return static_cast<int16_t>(_value);
 	}
-	constexpr float i32()const{
+	constexpr int32_t i32()const{
 		return static_cast<int32_t>(_value);
 	}
-	constexpr float u8()const{
+	constexpr uint8_t u8()const{
 		return static_cast<uint8_t>(_value);
 	}
-	constexpr float u16()const{
+	constexpr uint16_t u16()const{
 		return static_cast<uint16_t>(_value);
 	}
-	constexpr float u32()const{
+	constexpr uint32_t u32()const{
 		return static_cast<uint32_t>(_value);
 	}
